cheri: Return properly bounded pointers from gen_alloc_*()
Adjust the generic allocator in lib/genalloc.c and its callers to
return properly bounded pointers. The allocator is used by some
callers to manage a range of things that is not virtual memory and
thus does not have to be a valid pointer. Check what the caller
provides and only set bounds if we are in fact dealing with pointers.

+++ arch/arm/mach-davinci/sram.c
@@ -64,7 +64,7 @@ static int __init sram_init(void)
 		addr = ioremap(phys, len);
 		if (!addr)
 			return -ENOMEM;
-		status = gen_pool_add_virt(sram_pool, (unsigned long) addr,
+		status = gen_pool_add_virt(sram_pool, (uintptr_t) addr,
 					   phys, len, -1);
 		if (status < 0)
 			iounmap(addr);
